requirements.txt
appnope==0.1.4
asgiref==3.8.1
astropy==7.0.0
astropy-iers-data==0.2025.3.3.0.34.45
astroquery==0.4.9.post1
asttokens==3.0.0
beautifulsoup4==4.13.3
Brotli==1.0.9
certifi==2025.1.31
charset-normalizer==3.4.1
comm==0.2.2
contourpy==1.3.1
cycler==0.11.0
debugpy==1.8.11
decorator==5.2.1
Django==5.1.6
django-cors-headers==4.7.0
djangorestframework==3.15.2
et_xmlfile==2.0.0
exceptiongroup==1.2.2
executing==2.1.0
fonttools==4.55.3
html5lib==1.1
idna==3.10
importlib_metadata==8.6.1
ipykernel==6.29.5
ipython==9.0.0
ipython_pygments_lexers==1.1.1
jaraco.classes==3.4.0
jaraco.context==6.0.1
jaraco.functools==4.1.0
jax==0.4.35
jaxlib==0.4.35
jedi==0.19.2
joblib==1.4.2
jupyter_client==8.6.3
jupyter_core==5.7.2
jwxml==0.3.2
keyring==25.6.0
kiwisolver==1.4.8
lxml==5.3.1
matplotlib==3.10.0
matplotlib-inline==0.1.7
ml_dtypes==0.5.0
more-itertools==10.6.0
nest_asyncio==1.6.0
numexpr==2.10.2
numpy==2.2.2
openpyxl==3.1.5
opt-einsum==3.3.0
packaging==24.2
parso==0.8.4
pathlib==1.0.1
pexpect==4.9.0
photutils==2.2.0
pickleshare==0.7.5
pillow==11.1.0
pip==25.0
platformdirs==4.3.6
poppy==1.1.2
prompt_toolkit==3.0.50
psutil==7.0.0
psycopg2==2.9.9
ptyprocess==0.7.0
pure_eval==0.2.3
pyerfa==2.0.1.5
Pygments==2.19.1
pyparsing==3.2.0
pysiaf==0.24.1
python-dateutil==2.9.0.post0
pytz==2024.1
pyvo==1.6.1
PyYAML==6.0.2
pyzmq==26.2.0
requests==2.32.3
scipy==1.15.1
setuptools==72.1.0
six==1.16.0
soupsieve==2.6
sqlparse==0.5.2
stack_data==0.6.3
stpsf==2.0.0
stsynphot==1.4.0
synphot==1.5.0
tornado==6.4.2
tqdm==4.67.1
traitlets==5.14.3
typing_extensions==4.12.2
urllib3==2.3.0
wcwidth==0.2.13
webbpsf==1.5.0
webbpsf-ext==1.2.3
webencodings==0.5.1
wheel==0.45.1
zipp==3.21.0
